Recipe and Category fully work

+++ src/main/kotlin/com/example/recipes/models/FoodCategoryEntity.kt
@@ -8,5 +8,5 @@ data class FoodCategoryEntity(
 
     val category: String = "",
     val recipe_id: Int? = null,
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY) val food_category_id: Long? = null
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY) val food_category_id: Int? = null
 )
